Device specific apps to be translated
I added all device specific apps to the cancro folder, because cancro
had the most of them, when they are translated they can be copied to
aries and mocha. Please change values foldername to values-nl when
translating. when translation is complete the values folder can be
deleted and only the values-nl containing the translated files remain